GPencil: Refactor: Add stroke fast drawing
This use the same behavior as previous implementation. It uses a temp
buffer to store the screen and avoid redraw.

But this will be only on demand as the drawing may not match the final
look when releasing the pen. The UI option has not been added yet.

This mode still supports blending and VFX. Just stroke ordering with the
scene not supported.