GPencil: Refactor: Add SMAA as Antialiasing solution
This removes the need of MSAA and the cost associated with it.
We run edge detection on both revealage and color buffer.

However the implemntation is not perfect: The source buffers are in linear
color space and the edge detection implies gamma corrected (display) space.
This mean the algorithm fails to capture some color changes. To workaround
this, we increase SMAA_LUMA_WEIGHT which has the same effect as decreasing
SMAA_THRESHOLD. This works ok in most scenes since GPencil objects have
clean and sharply defined shapes.