GPencil: Move opacity factor to stroke from runtime
If the modifier is applied, need to be in the stroke.

Note: Maybe we need to active this value in some sculpt brush to define the opacity of the Fill, because when you apply the modiifer, you can only replace the value using the modifier again. Other solution would be a reset opacity operator.