OCIO: Use GroupTransform to avoid doing 3 transforms instead of 2
This exposes OCIO::GroupTransform to C api to be able to chain the
display transform with a corrective transform to output display linear
colors.

For now the display code only use an exponent transform to go back to
linear but should use a ColorSpaceTransform once we have the correct
config.