GPencil: Moves buffer drawing data to eval data
Instead to put the sbuffer data in the original datablock, now is copied to evaluated data, but without tagging all datablock that makes things very slow in fast drawing.

This changes removes the hacks to use the original datablock for buffers in the drawing engine.

Note: still there is a problem when a datablock is used several times with modifiers.