Support mirroring the VR view in regular 3D Views
Adds a per 3D View option to mirror the 3D session view. When enabled,
the viewport (continuously) renders using the same location and rotation
as the 3D view. While navigation is disabled then, other interactions
are working, so you can select and edit data as usual.
To be precise, we use the pose of the centroid, not the ones used by the
eyes.

Extra attention was put into forbidding viewport navigation while the VR
mirror runs (it overrides the current viewpoint anyway). Tried to make
this a generic feature. There may still be some navigation operators
that are not covered. So using them while the mirror runs will have an
affect on what you see once you end the session or mirroring.

Quadview is preserved, only the un-locked view shows the mirror then.
All of this is runtime data, no state changes should end up being read
from saved files.