Test runner: On Linux, use Python from the precompiled libraries
CentOS on the buildbot still runs Python 3.6, which is also used for the
unit tests. This means that the tests can't use language features that are
available to Blender itself. And testing with a different version of Python
than will be used by the actual code seems like a bad idea to me.