Tests: On Linux, use Python from install dir
CentOS on the buildbot still runs Python 3.6, which is also used for the
unit tests. This means that the tests can't use language features that
are available to Blender itself. And testing with a different version of
Python than will be used by the actual code seems like a bad idea to me.

This commit tries to find Python in the installation directory, which
will not be there on the first time CMake is run, so it also tries to
find it in the precompiled libraries. If that also fails, it defaults to
whichever Python executable CMake finds.

Differential Revision: https://developer.blender.org/D7395