Fix instanced panel list order when loading existing files
This just requires a small change to solve this situation: the panel
sort order needs to be set to the max sort order when it's added.

We recently got rid of similar behavior, but it turns out it's necessary
when loading existing files, and it doesn't hurt in other cases.