Outliner: Draw left column for activation/mode toggle
In Scenes view mode, a column is added when the toggle is enabled. In
view layer mode, the scene collection has a space next to it where a
expand/collapse triangle could be drawn, so this also offsets the tree a
unit to the left when in view layer display mode.