Outliner: Add buttons to set active camera
Add buttons to the left column to set the active camera. The active
camera currently shows a check, and other cameras draw a dot.

For some reason the dot doesn't have a proper hitbox area.

Right now the function to set the camera is done from button callbacks,
but this could also easily be done from outliner_select.c.