Outliner: Allow setting active data from column
Add a function to set active data from the outliner column on click. The
supported types are collections, cameras, and scenes as these types are
more "disruptive" and change the state of the scene.

For now this is set from the (cluttered) outliner selection function. I
think it would be best to attach to the drawn UI buttons, but currently
they don't want to work with attached functions.