Outliner: Draw icons for edit mode
The end goal is to support all interaction modes, but since edit mode
was already supported in the outliner, it makes sense to start here.

This code takes the current active object and determines it's mode. If
it is in edit mode, then the icons are drawn. This only works for
objects of the same type for now.