Outliner: Add edit mode toggle from left column
When in edit mode, compatible objects are shown next to a dot icon in
the outliner. These objects can then be toggled into the current edit
mode. When the active object is toggled out of edit mode, all objects
are removed from edit mode.

This currently has a few issues:
* Selection is not synced. Adding an object to edit mode selects it, and
  removing deselects it. This needs to be reflected in the outliner.
* This works for all types that support edit mode with two
  issues/exceptions:
** Grease Pencil edit mode is not supported (perhaps grease pencil needs
a small update)
** Text objects support multi-edit, but it doesn't actually allow you to
edit both objects.

The issues need to be fixed in the related mode-code, not the outliner.
It would be nice to fix so the outliner doesn't have to have cases to
not allow Text and Grease Pencil mode.