Outliner: Add walk navigation left and right
Support walk navigation up and down the tree by pressing the left and
right arrows. When holding shift (to toggle expand/collapse on all
children) walking is never done.

When walking right, open the tree if it is closed, or walk into the tree
if it is opened. When walking left, close the tree if it is open, or
select the parent element when closed.