Outliner: Show icons for all interaction modes
When not in object mode, show the interaction mode next to the object in
the outliner. Also show a dot next to each icon that could be switched
into the current mode.

Currently this still only works for edit and pose mode. Seeing the icon
next to the object in the mode is already useful. One issue is that
grease pencil mode icon draws green.