Outliner: Add mode switching to remaining modes in left column
When in modes other than edit or pose, icons are drawn next to each
object in the outliner that also supports the same mode. When the icon
is clicked, that object is activated and swapped into that edit mode.

When Lock Object Modes is disabled and objects are left in their modes,
the icons all draw as if they were simultaneously in that mode. This is
the only issue currently, otherwise it seems to work well.