Outliner: Fix mode column when lock object modes is disabled
When Lock Object Modes was disabled, multiple objects would be shown as
in the interaction mode even though only one object could be edited at a
time. This was not an issue for edit and pose mode because those modes
support multi object editing. But in modes like sculpt it implied that
multiple objects could be edited at the same time.

This only draws the active icon for the active object when lock object
modes is disabled, and when there is no active edit or pose object.