Outliner: Properties sync with pinned data
Previously only the type of tab/context was passed to the properties
editor from the outliner. This caused some issues. If Object A was
pinned, and Object B's mesh data was clicked, it would switch to Object
A's mesh data.

Now pass an RNA pointer along to verify if the outliner data is in the
valid path when the properties editor is pinned.

Also support Hair and PointCloud new object types.