Changed calulation of theta_u in tree importance.
This reduces artifacts when using the light tree without splitting.

Before the theta_u was calclulated using a bounding sphere to
approximate a bounding box. This sometimes resulted in the cone
containng the bounding box much, much too big. Now theta_u is found
more carefully so that it is the smallest angle such that the cone
completely contains the bounding box.