Outliner: Draw collection hierarchy lines
Rather than drawing lines for object hierarchies, draw for collections.
These lines will be colored when collections are colored.

There are still small issues with the implementation, but it works quite
well for now.

The dashed lines for child objects (not in collection) are removed but
could be brought back.