Outliner: Add generic UI Stack drop operator
Replaces the modifier drop operator and works for modifiers,
constraints, and shader fx. Because these three elements share common
traits, it made the code simpler to combine the operator for all three.
Three types of actions are supported depending on where the element is
dropped:
* Drop within the same list to reorder.
* Drop on another object to copy.
* Drop the base (parent) element to link all to another object.

A few cases are yet unsupported but will be added to blenkernel.