Outliner: Remove outliner parent object selection
In order to fix T74332 it was necessary to flag parent elements as
selected in the outliner to prevent the bases from being deselected on
the selection sync. This ensures the parent object is selected for
editing drivers, etc.

An annoying issue that results from this is the context menu shows
Object entries even when the target is a data (material, action, bone)
type. This also causes issues with drag and drop and other operations.

This adds a new TSE flag to prevent a selection sync on a per-element
case. This is used to prevent a parent element sync which ensures the
parent object stays selected, but it does not draw as selected in the
outliner.