Fix: Outliner small select sync issue with dragdrop
Dragdrop in the outliner would sync from objects rather than to objects.
This led to strange results when dragging child data. i.e. a modifier
drag would select the parent object in the outliner.

This requires not storing the dragged tree element because the tree is
rebuilt between the initial drag and the drop, causing heap use after
free.

Perhaps the outliner can be flagged to not rebuild on selection events.