Outliner: Infer outliner context menu from target element
The context menu tried to determine the type of action based on the
selected elements. This isn't intuitive, rather it should be based on
the target. That way if you have objects and collections selected, a
right click on a collection should show the collection menu.

There is more work to be done, some menus fail to execute properly now
that the old guarantees are gone.