Cleanup: Remove broken context menu mode toggle code
The outliner context menu had an operator to enter/exit modes, but it
only showed when the objects were in modes. Additionally, I'm not
certain how much it was used since it has been broken for a long time
and no one has reported it.

Entering modes through the context menu could be a nice feature, but I
think it would be an easier problem to solve after a context menu
refactor.