GPencil: Scale strokes to Potrace lib units
After doing a lot of testing, it looks the library is using a factor base on the VGA resolution and aspect ratio.

The formula was obtained experimenting and gets a very good result, but it could be a totally different approach if we find what logic is used by Potrace.